ChernobylVR_Demo version
If you want to play the demo version of the Chernobyl application in Virtual Reality. You need to download all the zip files from the github to a separate folder on your computer and then open to the zip archive "Demo_ChernobylVR_Unity3D.zip" and open the "ChernobylVR.exe" file.